[FEAT] restructure util.vis modules
