+++ pyrsa/vis/__init__.py
@@ -1,3 +1,3 @@
 from .rdm_plot import show_rdm
-from .vis import mds,rdm_dimension_reduction
+from .mds_plot import mds,rdm_dimension_reduction
 from .model_plot import plot_model_comparison
